requirements.txt
django
djangorestframework
Django>=5.2
psycopg[binary]
python-decouple
django-debug-toolbar
Faker
pytest
pytest-django